# src/main/java/uk/gov/dwp/health/pip/document/submission/manager/entity/DocumentId.java
package uk.gov.dwp.health.pip.document.submission.manager.entity;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class DocumentId {

  private String documentId;
}